+++ src/components/Calculate/CalculatePage.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import Button from "@material-ui/core/Button";
-import Expense from "../Expense";
+import Expense from "../Expense/Expense";
 import InputComponent from "../InputComponent/InputComponent";
 import Snackbar from "@material-ui/core/Snackbar";
 import { addParticipant, collectData } from "../../actions/index";
